+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [3.15.3] - 2024-06-10
 
+### Changed
+
+- Adjust autocomplete list to show product name rather then product sku name
+
 ## [3.15.2] - 2024-06-10
 
 ## [3.15.1] - 2024-05-27